% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Adduct_db}
\alias{Adduct_db}
\title{adduct data}
\format{
A \code{data.table} with the common adducts in LC-ESI-MS
\describe{
\item{adduct}{Formula}
\item{charge}{charge}
\item{mz_query}{Exact mass value}
}
}
\usage{
Adduct_db
}
\description{
adduct data
}
\keyword{datasets}
