% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{annotate_spectra}
\alias{annotate_spectra}
\title{Annotate a matrix and return network edges}
\usage{
annotate_spectra(mass, debugL = FALSE)
}
\arguments{
\item{mass}{A numeric vector}

\item{debugL}{logisfggdcal to print debug message}
}
\value{
Return a numeric vector of network edges
}
\description{
This function returns edges of networked
ions
}
