% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{CJ1}
\alias{CJ1}
\title{Cross join of two data.table}
\usage{
CJ1(DT1, DT2)
}
\arguments{
\item{DT1}{A data.table}

\item{DT2}{A data.table}
}
\value{
A data.table corresponding to the cross join of the two tables.
}
\description{
Cross join of two data.table
}
\examples{
require(data.table)
CJ1(
 data.table(A = 1:10, B = "A"),
 data.table(
   C = 50:100,
   D = rep(c("C", "D", "E", "F"), length.out = 51)
 )
)
}
