% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_monocharge}
\alias{db_monocharge}
\title{Mono charged ions}
\format{
A \code{data.table} with the common charges in LC-ESI-MS
\describe{
\item{Formula}{Formula}
\item{charge}{charge}
\item{lossL}{Does the resulting ion can produce in-source losses ?}
\item{mz_query}{Exact mass value}
}
}
\usage{
db_monocharge
}
\description{
Mono charged ions
}
\keyword{datasets}
