% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Element}
\alias{Element}
\title{Periodic table as a data.table}
\format{
A \code{data.table} with the Periodic Table Elements fields
\describe{
\item{atomic_nb}{Atomic number}
\item{atomic_symb}{Atomic symbol}
\item{mass_nb}{Mass number}
\item{atomic_mass}{Exact atomic mass}
\item{isotopic_compo}{Relative isotopic abundance}
}
}
\usage{
Element
}
\description{
Periodic table as a data.table
}
\references{
\url{https://physics.nist.gov/cgi-bin/Compositions/
stand_alone.pl?ele=&ascii=ascii2&isotype=some}
}
\keyword{datasets}
