% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_comb}
\alias{find_comb}
\title{Find all combination}
\usage{
find_comb(n, r)
}
\arguments{
\item{n}{number of pairs to create}

\item{r}{number of element in each pairs}
}
\description{
This function returns edges of networked
ions
}
