% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_space.R
\name{find_compo_from_mass}
\alias{find_compo_from_mass}
\title{Find composition from mass}
\usage{
find_compo_from_mass(
  mass_target,
  ppm = 5,
  use_golden_ratio = TRUE,
  elements_vc = NULL,
  debugl = 0
)
}
\arguments{
\item{mass_target}{mass to decompose}

\item{ppm}{mass error to filter the results (in ppm)}

\item{use_golden_ratio}{Logical to return the maximum
number of each element by using
the ratios from the seven golden
rules (\code{TRUE}) or not (\code{FALSE}).}

\item{elements_vc}{(optional) character vector containing
the element to include}

\item{debugl}{Integer (0: no message, 1: short, 2: verbose)}
}
\value{
A data.table with one proposition by line
with it's elemental composition, the theoretical
mass and the mass deviation with the query in
ppm.
}
\description{
Find composition from mass
}
\examples{
  find_compo_from_mass(125.215, ppm = 10)
  find_compo_from_mass(528.125, ppm = 3)
  find_compo_from_mass(89.0476, ppm = 3, elements_vc = "C3H7NO2")
}
