% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gen_adduct}
\alias{gen_adduct}
\title{Generate adduct list from mz}
\usage{
gen_adduct(
  mz,
  label = NULL,
  mz_type = c("neutral", "pos", "neg")[[1]],
  adduct_db = Spec2Annot::Adduct_db
)
}
\arguments{
\item{mz}{Exact mass}

\item{label}{(optional) a string to use in the ion labels
if not set it will be just 'X'.}

\item{mz_type}{Type of mass for targeted compound:
"pos", "neg" or "neutral"}

\item{adduct_db}{A data.table with adduct informations
see (\code{Spec2Annot::Adduct_db}).}
}
\value{
A data.table with adducts mass and labels.
}
\description{
Generate adduct list from mz
}
\examples{
gen_adduct(125.53658, "Isovalerine", "neutral")
gen_adduct(116.0837, "C6H12O2", "neutral")
gen_adduct(116.0837, "C6H12O2", "pos")
}
