% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gen_isotopes}
\alias{gen_isotopes}
\title{Function to generate isotopes from an exact mass}
\usage{
gen_isotopes(mz, label = NULL, db_iso = Spec2Annot::Isotopes_db)
}
\arguments{
\item{mz}{Exact mass}

\item{label}{(optional) a string to use in the ion labels
if not set it will be just 'X'.}

\item{db_iso}{A data.table with isotopes informations
(see \code{Spec2Annot::Isotopes_db} as example).}
}
\value{
Generate a data.table with isotopes mass and label
based on the \code{mz} value. By default, the isotopes are
chosen from \code{Spec2Annot::Isotopes_db} but it could
be specified by the user.
}
\description{
Function to generate isotopes from an exact mass
}
\examples{
gen_isotopes(125.53658, "Isovalerine")
}
