% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_space.R
\name{get_element_from_mass}
\alias{get_element_from_mass}
\title{Get Element from a mass}
\usage{
get_element_from_mass(mass, use_golden_ratio = FALSE)
}
\arguments{
\item{mass}{input mass}

\item{use_golden_ratio}{Logical to return the maximum
number of each element by using
the ratios from the seven golden
rules (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
Return a string with the number maximum number
of each element to expect from the mass.
get_element_from_mass(125.2535, FALSE)
get_element_from_mass(125.2535, TRUE)
}
\description{
Get Element from a mass
}
