% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_range}
\alias{get_range}
\title{Get index in range}
\usage{
get_range(input, valA, valB)
}
\arguments{
\item{input}{a numeric vector}

\item{valA}{a numeric value for the lower bound}

\item{valB}{a numeric value for the higher bound}
}
\value{
An \code{integer} vector of closest indexes found
}
\description{
Get index in range
}
