% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_senior}
\alias{get_senior}
\title{Check Senior theorems}
\usage{
get_senior(element_dt, global = TRUE)
}
\arguments{
\item{element_dt}{Either a data.table as returned by
\code{Spec2Annot::element_from_formula()}
or a string as "C18H12O3P".}

\item{global}{Logical to return a unique value if all the
theorem are valid (\code{TRUE}) or a vector with
the result of each theorem (\code{FALSE}).}
}
\value{
If global is set to \code{TRUE}, return a unique logical value
corresponding to \code{TRUE} if all the theorems pass or \code{FALSE}
if any of them isn't.
If global is set to \code{FALSE}, return a named logical vector with
the result of each theorem.
}
\description{
Check the Senior theorems as described in Morikawa and
Newbold (2003) with the following:
i) the sum of valencies is an even number, or the total
number of atoms having odd valencies is even.
ii) the sum of valencies is greater than or equal to twice
the maximum valency.
iii) the sum of valencies is greater than or equal to twice
the number of atoms minus 1.
}
\examples{
get_senior("C6H12O3")
}
\references{
\enumerate{
\item Senior JK (1951) Partitions and Their Representative Graphs. Am J Math 73:663. doi: 10.2307/2372318
\item Morikawa T, Newbold BT (2003) Analogous odd-even parities in mathematics and chemistry. Chemistry 12:445–450
}
}
