% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ions_database}
\alias{ions_database}
\title{ions data}
\format{
A \code{data.table} with the common ions found in LC-ESI-MS
\describe{
\item{Attribution}{Encoded formula}
\item{exact_mass}{Exact mass formula as a character string}
\item{charge}{charge}
\item{Type}{Molecular form (Monocharge, Loss, Adduct, ...)}
}
}
\usage{
ions_database
}
\description{
ions data
}
\keyword{datasets}
