% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Isotopes_db}
\alias{Isotopes_db}
\title{isotopes data}
\format{
A \code{data.table} with the common isotopes found in LC-ESI-MS
\describe{
\item{isotope}{Isotope formula}
\item{mass}{Exact mass}
\item{abundance}{Natural abundance (in percent)}
\item{mass_diff}{Exact mass difference from non-isotopic element}
}
}
\usage{
Isotopes_db
}
\description{
isotopes data
}
\keyword{datasets}
