% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Losses_db}
\alias{Losses_db}
\title{losses data}
\format{
A \code{data.table} with the common losses found in LC-ESI-MS
\describe{
\item{loss}{loss formula}
\item{mz_query}{Exact mass}
\item{encod}{Character encoding}
}
}
\usage{
Losses_db
}
\description{
losses data
}
\keyword{datasets}
