% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Monocharge_db}
\alias{Monocharge_db}
\title{Mono charged ions}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 4 columns.
}
\usage{
Monocharge_db
}
\description{
Mono charged ions
}
\keyword{datasets}
