% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{mz_from_string}
\alias{mz_from_string}
\title{Calculate mz from a string with signs}
\usage{
mz_from_string(string)
}
\arguments{
\item{string}{Formula as a string of the form "C6H5O3+" or with isotopes
\verb{[C6H5O3+H]+_13C1}}
}
\value{
Return a numeric value corresponding to the mass of the \code{string} input
}
\description{
Calculate mz from a string with signs
}
\examples{
mz_from_string("C6H5O3+")
mz_from_string("C6H5O3++")
mz_from_string("[C6H4O3+H]+_13C1")
mz_from_string("[C6H4O3+H]+_13C2")
}
