% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mz_ppm}
\alias{mz_ppm}
\title{Get mass deviation in ppm}
\usage{
mz_ppm(massa = 120.1253, massb = 120.1263)
}
\arguments{
\item{massa}{First mass}

\item{massb}{Second mass}
}
\value{
ppm as a \code{numeric} value between \code{massa} and \code{massb}.
}
\description{
This function returns a numeric value
corresponding to the mass deviation
between \code{massa} and \code{massb} in ppm.
}
