% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_join.R
\name{search_db}
\alias{search_db}
\title{Search DB egaint EXP}
\usage{
search_db(db_dt, exp_dt, ppmtol = 5, rttol = 5)
}
\arguments{
\item{db_dt}{a data.table with id, mz, rt columns}

\item{exp_dt}{a data.table with id, mz, rt columns}

\item{ppmtol}{numeric value to set ppm search in m/Z dimension.
Can be set to NULL to use only rt search.}

\item{rttol}{numeric value to set rt tolerance in rt dimension.
Can be set to NULL to use only mz search.}
}
\value{
Return the exp_dt data.table entries matching the criteria with
2 new columns: dbid (ids in db_dt) and expid (ids in exp_dt).
}
\description{
Search DB egaint EXP
}
