% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_join.R
\name{search_db_cpp}
\alias{search_db_cpp}
\title{Search DB egaint EXP using cpp function}
\usage{
search_db_cpp(in_db, in_exp, ppmtol = 5, rttol = 5)
}
\arguments{
\item{in_db}{a data.table with id, mz, rt columns}

\item{in_exp}{a data.table with id, mz, rt columns}

\item{ppmtol}{numeric value to set ppm search in m/Z dimension.}

\item{rttol}{numeric value to set rt tolerance in rt dimension.}
}
\value{
Return the exp_dt data.table entries matching the criteria with
2 new columns: dbid (ids in db_dt) and expid (ids in exp_dt).
}
\description{
Search DB egaint EXP using cpp function
}
