% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spec2Annot-package.R
\docType{package}
\name{Spec2Annot-package}
\alias{Spec2Annot}
\alias{Spec2Annot-package}
\title{Spec2Annot: Annotation of Mass Spectra}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a comprehensive suite of functions to efficiently annotate mass spectra data. Motivated by the need for rapid and accurate chemical identification in high-resolution mass spectrometry, it integrates built-in chemical databases and high-performance C++ algorithms. Users can perform mass-to-charge (m/Z) and retention time searches, determine elemental compositions of molecules using heuristic rules, including specific isotopes, and annotate MS2 spectra with structural metrics using configurable chemistry rules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odisce/Spec2Annot}
  \item Report bugs at \url{https://github.com/odisce/Spec2Annot/issues}
}

}
\author{
\strong{Maintainer}: Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr}

Authors:
\itemize{
  \item Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr}
}

Other contributors:
\itemize{
  \item Etienne Thevenot \email{etienne.thevenot@cea.fr} [contributor]
  \item Eric Venot \email{eric.venot@cea.fr} [reviewer]
  \item Annelaure Damont \email{annelaure.damont@cea.fr} [contributor]
  \item Anaïs Legrand \email{anais.legrand@cea.fr} [contributor]
}

}
\keyword{internal}
