% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectra_full}
\alias{spectra_full}
\title{Spectra example}
\format{
A \code{data.table} with a centroided MS spectra used as example
\describe{
\item{mz}{m/Z value}
\item{intensity}{Measured intensity}
\item{ID}{Unique identifier}
}
}
\usage{
spectra_full
}
\description{
Spectra example
}
\keyword{datasets}
