% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectra_ms2}
\alias{spectra_ms2}
\title{Spectra example}
\format{
A \code{data.table} with a centroided MS2 spectra used as example
\describe{
\item{mz}{m/Z value}
\item{i}{Measured intensity}
}
}
\usage{
spectra_ms2
}
\description{
MS2 spectrum of C10H12N5O6P1
Precursor mass: 330.0597
Mode: ESI +
Collision mode: HCD
Energy: 30
}
\keyword{datasets}
