% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valence_db}
\alias{valence_db}
\title{Valence data}
\format{
A \code{data.table} with element valence
\describe{
\item{atomic_symb}{Atomic symbol}
\item{valence}{valence value}
}
}
\usage{
valence_db
}
\description{
Valence data
}
\keyword{datasets}
