% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mz.R
\name{add_formula_to_annot}
\alias{add_formula_to_annot}
\title{Add formula to annotate_mz()}
\usage{
add_formula_to_annot(mzannot_dt)
}
\arguments{
\item{mzannot_dt}{A \code{data.table} containing the elemental composition of
ions as returned by \code{annotate_mz()}.}
}
\value{
Return the mzannot_dt table with a new \code{formula} column
containing a string with the elemental composition.
}
\description{
Add formula to annotate_mz()
}
