% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Adduct_db}
\alias{Adduct_db}
\title{adduct data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
Adduct_db
}
\description{
adduct data
}
\keyword{datasets}
