% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mz.R
\name{annotate_mz}
\alias{annotate_mz}
\title{Annotate a mass spectrum}
\usage{
annotate_mz(
  input_spectrum,
  ppm = 5,
  polarity = NULL,
  compo = NULL,
  use_golden_ratio = TRUE
)
}
\arguments{
\item{input_spectrum}{A mass spectrum as a data.table with \code{mz} and
\code{i } columns}

\item{ppm}{mass tolerance for the search in ppm.}

\item{polarity}{Polarity is used to convert the mass spectrum to neutral mass
before running the search.}

\item{compo}{(optional) Elemental composition as a string (\code{"C6H12O3"}) of
the neutral form.}

\item{use_golden_ratio}{Logical to return the maximum
number of each element by using
the ratios from the seven golden
rules (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
Return a data.table with the annotated spectrum.
}
\description{
This function annotate a full mass spectrum using a full brute force
search optionnaly restricted by an elemental composition \code{compo}.
Since generating the space to research can take some times, the
parameter \code{search_space} can be pre-calculated in advance.
}
\examples{
annotate_mz(spectra_ms2, ppm = 3, polarity = 1, compo = "C10H12N5O6P1")
}
