% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{annotate_spectra}
\alias{annotate_spectra}
\title{Annotate a matrix and return network edges}
\usage{
annotate_spectra(mass)
}
\arguments{
\item{mass}{A numeric vector}
}
\description{
This function returns edges of networked
ions
}
