% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{brute_force_const}
\alias{brute_force_const}
\title{Group m/Z from a vector based on tolerance}
\usage{
brute_force_const(
  mass = 120,
  ppm = 5,
  mass_vc = 0L,
  maxiter_vc_ = NULL,
  name_vc = 0L,
  debugl = 0L,
  debugit = 0L
)
}
\arguments{
\item{mass}{Targeted mass}

\item{ppm}{Mass tolerance in ppm to restrict results}

\item{mass_vc}{Vector of elemnt masses (size = n)}

\item{maxiter_vc_}{Vector of element maximum limits (optional) (size = n)}

\item{name_vc}{Vector of elemnt names (size = n)}

\item{debugl}{Integer (0: no message, 1: short, 2: verbose)}

\item{debugit}{Integer for max iter to do}
}
\description{
This function returns a matrix with m/Z groups
based on a tolerance.
}
