% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{electron_mass}
\alias{electron_mass}
\title{Return the mass of an electorn}
\usage{
electron_mass()
}
\value{
Return the mass of an electron as
a numeric value
}
\description{
Return the mass of an electorn
}
\examples{
electron_mass()
}
