% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Element}
\alias{Element}
\title{Periodic table as a data.table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 354 rows and 5 columns.
}
\usage{
Element
}
\description{
Periodic table as a data.table
}
\references{
\url{https://physics.nist.gov/cgi-bin/Compositions/
stand_alone.pl?ele=&ascii=ascii2&isotype=some}
}
\keyword{datasets}
