% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{element_from_formula}
\alias{element_from_formula}
\title{Get element count from formula}
\usage{
element_from_formula(formula)
}
\arguments{
\item{formula}{Formula as returned by \code{gen_formula_from_compo}}
}
\value{
Return a data.table with element count
}
\description{
Get element count from formula
}
\examples{
formula <- gen_formula_from_compo("C6H2O3Ca2K1")
element_from_formula(formula)
formula <- gen_formula_from_compo("[C6H12O2+H]+_13C2")
element_from_formula(formula)
}
