% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_ion_list.R
\name{fun_generate_ions_from_mz}
\alias{fun_generate_ions_from_mz}
\title{Calculate the ions database}
\usage{
fun_generate_ions_from_mz(
  mass = 252.2534,
  losses = Spec2Annot::Losses_db,
  iso = Spec2Annot::Isotopes_db,
  charges = Spec2Annot::db_monocharge,
  adducts = Spec2Annot::Adduct_db
)
}
\arguments{
\item{mass}{numerical mass to use as base}

\item{losses}{losses as data.table (see Spec2Annot::Losses_db)}

\item{charges}{charges as data.table (see Spec2Annot::db_monocharge)}

\item{adducts}{adducts as data.table (see Spec2Annot::Adduct_db)}
}
\value{
A data.table containing the full list of possible ions with the formula to calculate them
}
\description{
Calculate the ions database
}
\examples{
fun_get_ion_list()
}
