% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gen_losses}
\alias{gen_losses}
\title{Generate losses list from mz}
\usage{
gen_losses(
  mz,
  label = NULL,
  mz_type = c("neutral", "pos", "neg")[[1]],
  loss_db = Spec2Annot::Losses_db
)
}
\arguments{
\item{mz}{Exact mass}

\item{label}{(optional) a string to use in the ion labels
if not set it will be just 'X'.}

\item{mz_type}{Type of mass for targeted compound:
"pos", "neg" or "neutral"}

\item{loss_db}{A data.table with losses information
see \code{Spec2Annot::Losses_db}}
}
\value{
A data.table with losses mass and labels.
}
\description{
Generate losses list from mz
}
\examples{
gen_losses(125.53658, "Isovalerine", "neutral")
gen_losses(116.0837, "C6H12O2", "neutral")
gen_losses(116.0837, "C6H12O2", "pos")
}
