% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gen_monocharge}
\alias{gen_monocharge}
\title{Function to generate monocharges from an exact mass}
\usage{
gen_monocharge(
  mz,
  label = NULL,
  mz_type = c("neutral", "pos", "neg")[[1]],
  ion_mode = c("pos", "neg")[[1]],
  mono_db = Spec2Annot::db_monocharge
)
}
\arguments{
\item{mz}{Exact mass}

\item{label}{(optional) a string to use in the ion labels
if not set it will be just 'X'.}

\item{mz_type}{Type of mass for targeted compound:
"pos", "neg" or "neutral"}

\item{ion_mode}{Output mode wanted: "pos", "neg"}

\item{mono_db}{A data.table with charges information
(see \code{Spec2Annot::db_monocharge})}
}
\value{
A data.table with ID, label and mz_query
}
\description{
Function to generate monocharges from an exact mass
}
\examples{
gen_monocharge(125.53658, "Isovalerine", "neutral", "neg")
gen_monocharge(116.0837, "C6H12O2", "neutral", "neg")
gen_monocharge(116.0837, "C6H12O2", "neutral", "pos")
}
