% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_charge_from_compo}
\alias{get_charge_from_compo}
\title{Get charge from composition}
\usage{
get_charge_from_compo(compo)
}
\arguments{
\item{compo}{Elemental composition as a string (ex.: "C6H12O3NH2")}
}
\value{
Return the number of positive or negative charge
}
\description{
Get charge from composition
}
\examples{
get_charge_from_compo("C6H2O3NH4")
get_charge_from_compo("C6H2O3NH4+")
get_charge_from_compo("C6H2O3NH4++")
get_charge_from_compo("C6H2O3NH4-")
get_charge_from_compo("C6H2O3NH4---")
get_charge_from_compo("[C6H12O2+H-H2O]+_13C")
get_charge_from_compo("[2(C6H2O3)+NH4]++_13C3")
get_charge_from_compo("[2(C6H2O3)-NH4]--_13C3_15N1")
}
