% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_dbe}
\alias{get_dbe}
\title{Calculate DBE from an elemental composition}
\usage{
get_dbe(element_dt)
}
\arguments{
\item{element_dt}{Either a data.table as returned by
\code{Spec2Annot::element_from_formula()}
or a string as "C18H12O3P".}
}
\value{
Return a string from an element table
}
\description{
Calculation is performed as explained in
http://ms-textbook.com/chapter-6/answer-6-3/
}
\examples{
get_dbe("C32H25N3O2S")
}
