% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_nrule}
\alias{get_nrule}
\title{Check Nitrogen rule}
\usage{
get_nrule(n, mass)
}
\arguments{
\item{n}{Nitrogen atom number}

\item{mass}{Ion mass}
}
\value{
A logical corresponding to \code{TRUE} if the N
rule is respected or \code{FALSE} if not.
}
\description{
Calculation is performed as explained in
https://en.wikipedia.org/wiki/Nitrogen_rule
}
\examples{
get_nrule(5, 125.1235)
}
