% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ions_database}
\alias{ions_database}
\title{ions data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 958 rows and 6 columns.
}
\usage{
ions_database
}
\description{
ions data
}
\keyword{datasets}
