% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Isotopes_db}
\alias{Isotopes_db}
\title{isotopes data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 12 rows and 4 columns.
}
\usage{
Isotopes_db
}
\description{
isotopes data
}
\keyword{datasets}
