% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Losses_db}
\alias{Losses_db}
\title{losses data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 182 rows and 3 columns.
}
\usage{
Losses_db
}
\description{
losses data
}
\keyword{datasets}
