% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_tables}
\alias{match_tables}
\title{Match two table}
\usage{
match_tables(db_dt, exp_dt, ppmtol, rttol, debugL = FALSE)
}
\arguments{
\item{db_dt}{a data.frame with at least mz and rt values (ref table)}

\item{exp_dt}{a data.frame with at least mz and rt values (exp table)}

\item{ppmtol}{a numeric value for the ppm tolerance}

\item{rttol}{a numeric value for the rt tolerance}
}
\description{
This function match two table, searching every
entries in B which are contained in A mz +- ppmtol
and A.rt +- rttol
}
