% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{mz_calc_ion}
\alias{mz_calc_ion}
\title{Calculate ion mass using form}
\usage{
mz_calc_ion(mass, form = "-H")
}
\arguments{
\item{mass}{numerical mass to use as base}

\item{form}{chemical form to add or substract (on of \code{Spec2Annot::db_monocharge[, unique(Formula)]})}
}
\value{
A numeric value corresponding to the mass form.
}
\description{
Calculate ion mass using form
}
\examples{
mz_calc_ion(142.5236, "-H")
}
