% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{mz_ppm}
\alias{mz_ppm}
\title{Calculate ppm deviation between two masses}
\usage{
mz_ppm(massa, massb)
}
\arguments{
\item{massa}{First mass as \code{numeric()}}

\item{massb}{Second mass as \code{numeric()}}
}
\value{
A numeric value corresponding to the ppm deviation between massa and massb
}
\description{
Calculate ppm deviation between two masses
}
\examples{
mz_ppm(142.5236, 142.5241)
}
