% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{mz_range}
\alias{mz_range}
\title{Calculate mass range using a ppm deviation}
\usage{
mz_range(mass, ppm)
}
\arguments{
\item{mass}{mass to use as center value}

\item{ppm}{ppm deviation to get the mass range}
}
\value{
A numeric vector with the mass range corresponding to mass +- ppm/2.
The range difference is equal to the asked ppm
}
\description{
Calculate mass range using a ppm deviation
}
\examples{
mz_range(142.5236, 5)
}
