% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mz_range}
\alias{mz_range}
\title{Get mass range with ppm}
\usage{
mz_range(mass = 120.1253, ppm = 10)
}
\arguments{
\item{mass}{First mass}

\item{ppm}{ppm tolerance}
}
\description{
Get mass range with ppm
}
