% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mz_vec_aggregate}
\alias{mz_vec_aggregate}
\title{Group m/Z from a vector based on tolerance}
\usage{
mz_vec_aggregate(xx, tt)
}
\arguments{
\item{xx}{A numeric and sorted vector of m/Z}

\item{tt}{Tolerance in absolute m/Z to group peaks}
}
\description{
This function returns a matrix with m/Z groups
based on a tolerance.
}
