% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spec2Annot-package.R
\docType{package}
\name{Spec2Annot-package}
\alias{Spec2Annot}
\alias{Spec2Annot-package}
\title{Spec2Annot: Annotation of Mass Spectra}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This R package implement different way to annotate and query Mass spectra, using brute-force, isotopic or fragmentation pattern, prior informations on losses or adduct, etc...
}
\author{
\strong{Maintainer}: Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr}

Other contributors:
\itemize{
  \item Etienne Thevenot \email{etienne.thevenot@cea.fr} [contributor]
  \item Eric Venot \email{eric.venot@cea.fr} [reviewer]
  \item Annelaure Damont \email{annelaure.damont@cea.fr} [contributor]
  \item Anaïs Legrand \email{anais.legrand@cea.fr} [contributor]
}

}
\keyword{internal}
