% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectra_full}
\alias{spectra_full}
\title{Spectra exemple}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2484 rows and 3 columns.
}
\usage{
spectra_full
}
\description{
Spectra exemple
}
\keyword{datasets}
