% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{string_from_element}
\alias{string_from_element}
\title{Return a string from an element table}
\usage{
string_from_element(element_dt)
}
\arguments{
\item{element_dt}{A data.table as returned by
\code{Spec2Annot::element_from_formula()}}
}
\value{
Return a string from an element table
}
\description{
Return a string from an element table
}
\examples{
"C6H12O3" \%>\%
  gen_formula_from_compo() \%>\%
  element_from_formula() \%>\%
  string_from_element()
}
