% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valence_db}
\alias{valence_db}
\title{Valence data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 117 rows and 2 columns.
}
\usage{
valence_db
}
\description{
Valence data
}
\keyword{datasets}
